PixArt disconnect: Explain bit mask and reference usbhid
The GitHub traffic graph revealed that some Japanese guys had
issues with a Microsoft PixArt mouse. They only read the section
about the PixArt mouse disconnecting and did not know that 0x00000400
is the HID_QUIRK_ALWAYS_POLL. So mention that there again and
reference to the section about the usbhid driver.

Reference:
http://linet.gr.jp/~kojima/PlamoWeb/ML/htdocs/201808/msg00019.html
Fixes #10