Point to Logitech based mice in the wireless section
Not all wireless mice are really dangerous. But the majority is.
E.g. an Apple Magic Mouse 2 has been found with no measurable
radiation increase in an environment of 1 V/m. Apple is known for
radiation-aware development. So point to the majority which is
Logitech/PixArt based and with >= 6 V/m really harmful.